requirements.txt
Django==3.2.24
django-cors-headers==4.1.0
djangorestframework==3.14.0
phonenumbers==8.13.31
gunicorn==21.2.0
